Refactoring the Code into component style
